+++ src/GRA.Data/Model/DynamicAvatarElement.cs
@@ -5,12 +5,9 @@ namespace GRA.Data.Model
     public class DynamicAvatarElement : Abstract.BaseDbEntity
     {
         [Required]
-        public int DynamicAvatarLayerId { get; set; }
+        public int DynamicAvatarId { get; set; }
 
         [Required]
-        [MaxLength(255)]
-        public string Name { get; set; }
-        [Required]
-        public int Position { get; set; }
+        public int DynamicAvatarLayerId { get; set; }
     }
 }