+++ src/GRA.Domain.Model/DynamicAvatar.cs
@@ -1,0 +1,16 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.Collections.Generic;
+
+namespace GRA.Domain.Model
+{
+    public class DynamicAvatar : Abstract.BaseDomainEntity
+    {
+        [Required]
+        [MaxLength(255)]
+        public string Name { get; set; }
+
+        [Required]
+        public int Position { get; set; }
+        public IEnumerable<DynamicAvatarElement> Elements { get; set; }
+    }
+}