+++ src/GRA.Domain.Repository/IDynamicAvatarElementRepository.cs
@@ -7,7 +7,6 @@ namespace GRA.Domain.Repository
     {
         Task<bool> ExistsAsync(int dynamicAvatarLayerId, int id);
         Task<int> GetFirstElement(int dynamicAvatarLayerId);
-        Task<int> GetIdByLayerIdAsync(int dynamicAvatarLayerId);
         Task<int> GetLastElement(int dynamicAvatarLayerId);
         Task<int?> GetNextElement(int dynamicAvatarLayerId, int elementId);
         Task<int?> GetPreviousElement(int dynamicAvatarLayerId, int elementId);